refactor: :truck: update e2e fixture file structure
