--- test/react-native/features/fixtures/performance-scenarios/index.js
@@ -1,1 +1,0 @@
-export { launchScenario } from './ScenarioLauncher'
