+++ test/react-native/features/fixtures/scenario-launcher/lib/ConfigFileReader.js
