+++ test/react-native/features/fixtures/scenario-launcher/package.json
