+++ test/react-native/features/fixtures/scenario-launcher/scenarios/AppStartScenario.js
